requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==1.0.3
telethon==1.28.5
langchain-groq==0.0.1
httpx
python-logging-loki==0.3.1
gunicorn==21.2.0
fastapi==0.104.1
uvicorn==0.24.0